feat: [~] move sorting functions to 3.x and other tests to correct file
